Introduce didactic default natlink.ini file for test of installation This commit completes the default search for a natlink.ini file, which has been renamed here as well:
   .natlink -> natlink.ini

to help ordinary Windows users (so file is not hidden).  A new directory InstallTest now comes with
an installation of natlink. The natlink.ini file there is being read, if none is found in "~/". It loads
a couple of module stubs and the demo is live: call backs etc. are shown because this natlink.ini
sets the debug level to, well, DEBUG!

CMAKE files have not been updated with new dependcies -- will happen next.