Make some changes to Natlink's message window (#214)
* Fix Natlink message window pop-up menu initialization

Re: #184.

The message window's pop-up menu is now loaded from the executable
resource data.  It is easier to change now.

* Add an Exit item to the Natlink message window's File sub-menu

Re: #213.

This item is grayed out (disabled) by default, for now.  I've added
a separator between it and the File>Reload item.

* Adjust File>Exit menu item changes to do things the C++/Windows way

Re: #184.

That is, use boolean flags and a dedicated user window message to
convey updates to the output window.  This makes it easier to update
the pop-up menu later down the line, if desired, by adding and hand-
ling new flags.

* Add a new function for controlling Natlink's message window

Re: #213.

The new function is setMessageWindow().  I have documented it in the
NatlinkSource/natlink.txt file.

It is now necessary to call this function with a Python callback to
enable the message window.  The default callback will soon reside in
the natlinkcore code and be set from appsupp.cpp.

Since it is related, this changeset includes modifications to the
message window's File>Reload logic, re: #28.  The default callback
will do a narrower reloading of user modules.

* Fix global callbacks that fail without thread safety

This changes the CDragonCode::makeCallback() method to hold Python's
GIL regardless of whether Natlink thread safety is enabled.  A crash
occurs if this is not done.

* Adjust and fix the message window changes a little

Re: #213.

* Update appsupp

- Show the message window before natConnect().
- Remove the now unused reloadPython() method.

* Fix some indentation issues in two files

* Fix a problem with setMessageWindow()

* Stop releasing natlink objects on IDD_RELOAD

This can cause inconsistent state in user/library code.

---------